Changed template

+++ 1.cpp
@@ -4,7 +4,7 @@ using namespace std;
 using ll = long long;
 using vi = vector<int>;
 using vll = vector<ll>;
-using pii = pair<int, int>;
+using pii = pair<ll, ll>;
 
 #define all(v) (v).begin(), (v).end()
 #define pb push_back
